Crate authzen_opa

Structs

AccountSessionFields

DeleteDocument

DeleteDocumentParams

GetConfig

GetConfigParams

GetDocument

GetDocumentParams

GetDocumentResult

GetStatus

GetStatusParams

Health

HealthParams

OPAClient

OPAPolicyASTNodeRef

OPAPolicyASTNodeRefRef

OPAQuery

OPAQueryConfig

OPAQueryInput

OPAQueryResult

OPAResult

OPATxEntity

OPATxEntityFull

OPAType

PatchDocument

uses json patch format

UpsertDocument

UpsertDocumentParams

Enums

CowSlice

DisableAuthorization

use this type in places where authorization disabling needs to be very clear in all contexts e.g. as a function parameter, where both the function and the function caller should be unambiguously aware of what the parameter they are using / passing means

OPAPolicyASTNode

OPAPolicyPathNode

OPAQueryInputAction

Constants

DEFAULT_TIMEOUT_SECONDS

DEFAULT_TTL_SECONDS

TTL_INDEX_NAME

Traits

AllOPAIdentifiable

AuthzService

AuthzServiceEntity

OPAContext

OPAIdentifiable

OPAIdentifiableEntity

OPAKey

OPATxCacheClient

OPATxCacheContext

Functions

initialize_ttl_index

query_all

Type Definitions

OPADataCacheUpsert

OPAMongoCollection

OPAMongoCollectionValue

Derive Macros

OPAContext

OPATxCacheContext

OPAType